+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdRequest.java
@@ -15,9 +15,11 @@
 package io.flutter.plugins.googlemobileads;
 
 import android.os.Bundle;
+import android.util.Pair;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import com.google.ads.mediation.admob.AdMobAdapter;
+import com.google.android.gms.ads.AbstractAdRequestBuilder;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.mediation.MediationExtrasReceiver;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
@@ -34,9 +36,14 @@ class FlutterAdRequest {
   @Nullable private final List<String> neighboringContentUrls;
   @Nullable private final Integer httpTimeoutMillis;
   @Nullable private final String mediationExtrasIdentifier;
-  @Nullable private final MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
+
+  @SuppressWarnings("deprecation")
+  @Nullable
+  private final MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
+
   @Nullable private final Map<String, String> adMobExtras;
   @NonNull private final String requestAgent;
+  @Nullable private final List<FlutterMediationExtras> mediationExtras;
 
   protected static class Builder {
     @Nullable private List<String> keywords;
@@ -45,9 +52,15 @@ class FlutterAdRequest {
     @Nullable private List<String> neighboringContentUrls;
     @Nullable private Integer httpTimeoutMillis;
     @Nullable private String mediationExtrasIdentifier;
-    @Nullable private MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
+
+    @SuppressWarnings("deprecation")
+    @Nullable
+    private MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
+
     @Nullable private Map<String, String> adMobExtras;
     @NonNull private String requestAgent;
+
+    @Nullable private List<FlutterMediationExtras> mediationExtras;
 
     @CanIgnoreReturnValue
     Builder setRequestAgent(String requestAgent) {
@@ -92,6 +105,7 @@ class FlutterAdRequest {
     }
 
     @CanIgnoreReturnValue
+    @SuppressWarnings("deprecation")
     Builder setMediationNetworkExtrasProvider(
         @Nullable MediationNetworkExtrasProvider mediationNetworkExtrasProvider) {
       this.mediationNetworkExtrasProvider = mediationNetworkExtrasProvider;
@@ -101,6 +115,12 @@ class FlutterAdRequest {
     @CanIgnoreReturnValue
     Builder setAdMobExtras(@Nullable Map<String, String> adMobExtras) {
       this.adMobExtras = adMobExtras;
+      return this;
+    }
+
+    @CanIgnoreReturnValue
+    Builder setMediationExtras(@Nullable List<FlutterMediationExtras> mediationExtras) {
+      this.mediationExtras = mediationExtras;
       return this;
     }
 
@@ -134,6 +154,7 @@ class FlutterAdRequest {
       return mediationExtrasIdentifier;
     }
 
+    @SuppressWarnings("deprecation")
     @Nullable
     protected MediationNetworkExtrasProvider getMediationNetworkExtrasProvider() {
       return mediationNetworkExtrasProvider;
@@ -149,6 +170,11 @@ class FlutterAdRequest {
       return requestAgent;
     }
 
+    @Nullable
+    protected List<FlutterMediationExtras> getMediationExtras() {
+      return mediationExtras;
+    }
+
     FlutterAdRequest build() {
       return new FlutterAdRequest(
           keywords,
@@ -159,10 +185,12 @@ class FlutterAdRequest {
           mediationExtrasIdentifier,
           mediationNetworkExtrasProvider,
           adMobExtras,
-          requestAgent);
+          requestAgent,
+          mediationExtras);
     }
   }
 
+  @SuppressWarnings("deprecation")
   protected FlutterAdRequest(
       @Nullable List<String> keywords,
       @Nullable String contentUrl,
@@ -172,7 +200,8 @@ class FlutterAdRequest {
       @Nullable String mediationExtrasIdentifier,
       @Nullable MediationNetworkExtrasProvider mediationNetworkExtrasProvider,
       @Nullable Map<String, String> adMobExtras,
-      String requestAgent) {
+      String requestAgent,
+      @Nullable List<FlutterMediationExtras> mediationExtras) {
     this.keywords = keywords;
     this.contentUrl = contentUrl;
     this.nonPersonalizedAds = nonPersonalizedAds;
@@ -182,12 +211,20 @@ class FlutterAdRequest {
     this.mediationNetworkExtrasProvider = mediationNetworkExtrasProvider;
     this.adMobExtras = adMobExtras;
     this.requestAgent = requestAgent;
+    this.mediationExtras = mediationExtras;
   }
 
   /** Adds network extras to the ad request builder, if any. */
-  private void addNetworkExtras(AdRequest.Builder builder, String adUnitId) {
+  private <T extends AbstractAdRequestBuilder<T>> void addNetworkExtras(
+      AbstractAdRequestBuilder<T> builder, String adUnitId) {
     Map<Class<? extends MediationExtrasReceiver>, Bundle> networkExtras = new HashMap<>();
-    if (mediationNetworkExtrasProvider != null) {
+    if (mediationExtras != null) {
+      for (FlutterMediationExtras flutterExtras : mediationExtras) {
+        Pair<Class<? extends MediationExtrasReceiver>, Bundle> pair =
+            flutterExtras.getMediationExtras();
+        networkExtras.put(pair.first, pair.second);
+      }
+    } else if (mediationNetworkExtrasProvider != null) {
       Map<Class<? extends MediationExtrasReceiver>, Bundle> providedExtras =
           mediationNetworkExtrasProvider.getMediationExtras(adUnitId, mediationExtrasIdentifier);
       networkExtras.putAll(providedExtras);
@@ -216,7 +253,9 @@ class FlutterAdRequest {
   }
 
   /** Updates the {@link AdRequest.Builder} with the properties in this {@link FlutterAdRequest}. */
-  protected AdRequest.Builder updateAdRequestBuilder(AdRequest.Builder builder, String adUnitId) {
+  protected <T extends AbstractAdRequestBuilder<T>>
+      AbstractAdRequestBuilder<T> updateAdRequestBuilder(
+          AbstractAdRequestBuilder<T> builder, String adUnitId) {
     if (keywords != null) {
       for (final String keyword : keywords) {
         builder.addKeyword(keyword);
@@ -237,7 +276,7 @@ class FlutterAdRequest {
   }
 
   AdRequest asAdRequest(String adUnitId) {
-    return updateAdRequestBuilder(new AdRequest.Builder(), adUnitId).build();
+    return ((AdRequest.Builder) updateAdRequestBuilder(new AdRequest.Builder(), adUnitId)).build();
   }
 
   @Nullable
@@ -280,6 +319,11 @@ class FlutterAdRequest {
     return requestAgent;
   }
 
+  @Nullable
+  protected List<FlutterMediationExtras> getMediationExtras() {
+    return mediationExtras;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -308,6 +352,7 @@ class FlutterAdRequest {
         neighboringContentUrls,
         httpTimeoutMillis,
         mediationExtrasIdentifier,
-        mediationNetworkExtrasProvider);
+        mediationNetworkExtrasProvider,
+        mediationExtras);
   }
 }