+++ packages/google_mobile_ads/example/ios/RunnerTests/FLTAppOpenAdTest.m
@@ -56,26 +56,19 @@
 // Helper method for testing with FLTAdRequest and FLTGAMAdRequest.
 - (void)testLoadShowAppOpenAd:(FLTAdRequest *)request
               gadOrGAMRequest:(GADRequest *)gadOrGAMRequest {
-  UIViewController *mockRootViewController =
-      OCMClassMock([UIViewController class]);
-  FLTAppOpenAd *ad =
-      [[FLTAppOpenAd alloc] initWithAdUnitId:@"testId"
-                                     request:request
-                          rootViewController:mockRootViewController
-                                 orientation:@1
-                                        adId:@1];
+  FLTAppOpenAd *ad = [[FLTAppOpenAd alloc] initWithAdUnitId:@"testId"
+                                                    request:request
+                                                       adId:@1];
   ad.manager = mockManager;
 
   // Stub the load call to invoke successful load callback.
   id appOpenClassMock = OCMClassMock([GADAppOpenAd class]);
-  OCMStub(ClassMethod([appOpenClassMock
-               loadWithAdUnitID:[OCMArg any]
-                        request:[OCMArg any]
-                    orientation:UIInterfaceOrientationPortrait
-              completionHandler:[OCMArg any]]))
+  OCMStub(ClassMethod([appOpenClassMock loadWithAdUnitID:[OCMArg any]
+                                                 request:[OCMArg any]
+                                       completionHandler:[OCMArg any]]))
       .andDo(^(NSInvocation *invocation) {
         void (^completionHandler)(GADAppOpenAd *ad, NSError *error);
-        [invocation getArgument:&completionHandler atIndex:5];
+        [invocation getArgument:&completionHandler atIndex:4];
         completionHandler(appOpenClassMock, nil);
       });
   // Stub setting of FullScreenContentDelegate to invoke delegate callbacks.
@@ -113,7 +106,6 @@
   OCMVerify(ClassMethod([appOpenClassMock
        loadWithAdUnitID:[OCMArg isEqual:@"testId"]
                 request:[OCMArg isEqual:gadOrGAMRequest]
-            orientation:UIInterfaceOrientationPortrait
       completionHandler:[OCMArg any]]));
   OCMVerify([mockManager onAdLoaded:[OCMArg isEqual:ad]
                        responseInfo:[OCMArg isEqual:responseInfo]]);
@@ -134,8 +126,7 @@
 
   [ad show];
 
-  OCMVerify([appOpenClassMock
-      presentFromRootViewController:[OCMArg isEqual:mockRootViewController]]);
+  OCMVerify([appOpenClassMock presentFromRootViewController:[OCMArg isNil]]);
 
   // Verify full screen callbacks.
   OCMVerify([mockManager adWillPresentFullScreenContent:[OCMArg isEqual:ad]]);
@@ -166,36 +157,27 @@
 
 // Helper for testing failed to load.
 - (void)testFailedToLoad:(FLTAdRequest *)request {
-  UIViewController *mockRootViewController =
-      OCMClassMock([UIViewController class]);
-  FLTAppOpenAd *ad =
-      [[FLTAppOpenAd alloc] initWithAdUnitId:@"testId"
-                                     request:request
-                          rootViewController:mockRootViewController
-                                 orientation:@2
-                                        adId:@1];
+  FLTAppOpenAd *ad = [[FLTAppOpenAd alloc] initWithAdUnitId:@"testId"
+                                                    request:request
+                                                       adId:@1];
   ad.manager = mockManager;
 
   id appOpenClassMock = OCMClassMock([GADAppOpenAd class]);
   NSError *error = OCMClassMock([NSError class]);
-  OCMStub(ClassMethod([appOpenClassMock
-               loadWithAdUnitID:[OCMArg any]
-                        request:[OCMArg any]
-                    orientation:UIInterfaceOrientationLandscapeLeft
-              completionHandler:[OCMArg any]]))
+  OCMStub(ClassMethod([appOpenClassMock loadWithAdUnitID:[OCMArg any]
+                                                 request:[OCMArg any]
+                                       completionHandler:[OCMArg any]]))
       .andDo(^(NSInvocation *invocation) {
         void (^completionHandler)(GADAppOpenAd *ad, NSError *error);
-        [invocation getArgument:&completionHandler atIndex:5];
+        [invocation getArgument:&completionHandler atIndex:4];
         completionHandler(nil, error);
       });
 
   [ad load];
 
-  OCMVerify(ClassMethod([appOpenClassMock
-       loadWithAdUnitID:[OCMArg any]
-                request:[OCMArg any]
-            orientation:UIInterfaceOrientationLandscapeLeft
-      completionHandler:[OCMArg any]]));
+  OCMVerify(ClassMethod([appOpenClassMock loadWithAdUnitID:[OCMArg any]
+                                                   request:[OCMArg any]
+                                         completionHandler:[OCMArg any]]));
   OCMVerify([mockManager onAdFailedToLoad:[OCMArg isEqual:ad]
                                     error:[OCMArg isEqual:error]]);
 }