+++ packages/google_mobile_ads/ios/Classes/FLTAd_Internal.h
@@ -14,6 +14,7 @@
 
 #import "FLTAdInstanceManager_Internal.h"
 #import "FLTGoogleMobileAdsPlugin.h"
+#import "FLTMediationExtras.h"
 #import "FLTMediationNetworkExtrasProvider.h"
 #import "FLTMobileAds_Internal.h"
 #import "FLTNativeTemplateStyle.h"
@@ -86,11 +87,14 @@
 @property NSString *_Nullable contentURL;
 @property BOOL nonPersonalizedAds;
 @property NSArray<NSString *> *_Nullable neighboringContentURLs;
-@property NSString *_Nullable mediationExtrasIdentifier;
+@property NSString *_Nullable mediationExtrasIdentifier
+    DEPRECATED_MSG_ATTRIBUTE("Use mediationExtras instead.");
 @property id<
-    FLTMediationNetworkExtrasProvider> _Nullable mediationNetworkExtrasProvider;
+    FLTMediationNetworkExtrasProvider> _Nullable mediationNetworkExtrasProvider
+    DEPRECATED_MSG_ATTRIBUTE("Use mediationExtras instead.");
 @property NSDictionary<NSString *, NSString *> *_Nullable adMobExtras;
 @property NSString *_Nonnull requestAgent;
+@property NSArray<id<FlutterMediationExtras>> *_Nullable mediationExtras;
 
 - (GADRequest *_Nonnull)asGADRequest:(NSString *_Nonnull)adUnitId;
 @end
@@ -210,8 +214,6 @@
 @interface FLTInterstitialAd : FLTFullScreenAd
 - (instancetype _Nonnull)initWithAdUnitId:(NSString *_Nonnull)adUnitId
                                   request:(FLTAdRequest *_Nonnull)request
-                       rootViewController:
-                           (UIViewController *_Nonnull)rootViewController
                                      adId:(NSNumber *_Nonnull)adId;
 - (GADInterstitialAd *_Nullable)interstitial;
 - (NSString *_Nonnull)adUnitId;
@@ -222,16 +224,12 @@
 @interface FLTGAMInterstitialAd : FLTInterstitialAd <GADAppEventDelegate>
 - (instancetype _Nonnull)initWithAdUnitId:(NSString *_Nonnull)adUnitId
                                   request:(FLTGAMAdRequest *_Nonnull)request
-                       rootViewController:
-                           (UIViewController *_Nonnull)rootViewController
                                      adId:(NSNumber *_Nonnull)adId;
 @end
 
 @interface FLTRewardedAd : FLTFullScreenAd
 - (instancetype _Nonnull)initWithAdUnitId:(NSString *_Nonnull)adUnitId
                                   request:(FLTAdRequest *_Nonnull)request
-                       rootViewController:
-                           (UIViewController *_Nonnull)rootViewController
                                      adId:(NSNumber *_Nonnull)adId;
 - (GADRewardedAd *_Nullable)rewardedAd;
 - (void)setServerSideVerificationOptions:
@@ -241,8 +239,6 @@
 @interface FLTRewardedInterstitialAd : FLTFullScreenAd
 - (instancetype _Nonnull)initWithAdUnitId:(NSString *_Nonnull)adUnitId
                                   request:(FLTAdRequest *_Nonnull)request
-                       rootViewController:
-                           (UIViewController *_Nonnull)rootViewController
                                      adId:(NSNumber *_Nonnull)adId;
 - (GADRewardedInterstitialAd *_Nullable)rewardedInterstitialAd;
 - (void)setServerSideVerificationOptions:
@@ -252,9 +248,6 @@
 @interface FLTAppOpenAd : FLTFullScreenAd
 - (instancetype _Nonnull)initWithAdUnitId:(NSString *_Nonnull)adUnitId
                                   request:(FLTAdRequest *_Nonnull)request
-                       rootViewController:
-                           (UIViewController *_Nonnull)rootViewController
-                              orientation:(NSNumber *_Nonnull)orientation
                                      adId:(NSNumber *_Nonnull)adId;
 - (GADAppOpenAd *_Nullable)appOpenAd;
 @end